Python/pyproject.toml
[project]
name = "python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "numpy>=2.4.2",
    "pandas>=3.0.1",
    "pillow>=12.1.1",
    "pycryptodome>=3.23.0",
    "pydantic>=2.12.5",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
